+++ CHANGELOG.md
@@ -10,6 +10,15 @@ internal API changes are not present.
 Main (unreleased)
 -----------------
 
+## Features
+
+- Introduce Apache HTTP exporter integration. (@v-zhuravlev)
+
+- Introduce eBPF exporter integration. (@tpaschalis)
+
+v0.25.0 (2022-06-06)
+-------------------------
+
 > **BREAKING CHANGES**: This release has breaking changes. Please read entries
 > carefully and consult the [upgrade guide][] for specific instructions.